refactor gdn
